removed master header. Now all files have required header files inside.
