+++ CMakeLists.txt
@@ -7,7 +7,7 @@ include_directories(${MPI_INCLUDE_PATH})
 find_package(HDF5 REQUIRED)
 set(HDF5_PREFER_PARALLEL TRUE)
 include_directories("/home/alcauchy/hdf5/hdf5/include")
-add_executable(Alliance_alpha_1_0 main.c master.h mpi_utils.h mpi_utils.c master.c fftw_utils.c fftw_utils.h hdf_utils.h hdf_utils.c parameters_io.c parameters_io.h diagnostics.h diagnostics.c)
+add_executable(Alliance_alpha_1_0 main.c mpi_utils.h mpi_utils.c fftw_utils.c fftw_utils.h hdf_utils.h hdf_utils.c parameters_io.c parameters_io.h diagnostics.h diagnostics.c array.h array.c)
 target_link_libraries(Alliance_alpha_1_0 "/home/alcauchy/hdf5/hdf5/lib/libhdf5.so")
 target_link_libraries(Alliance_alpha_1_0 "/usr/local/lib/libfftw3_mpi.a")
 target_link_libraries(Alliance_alpha_1_0 "/usr/local/lib/libfftw3.a")
